[Update] Reorganise types
