+++ components/tribes/players/IndividualPlayerView.tsx
@@ -17,9 +17,7 @@ import {
   Trophy,
   Target,
 } from "lucide-react";
-import { GameSession } from "@/components/tribes/TribeHomeTab";
-import { TribeMember } from "@/components/tribes/TribePlayersTab";
-import { HistStatsInterface } from "@/components/tribes/TribePageClient";
+import { type GameSession, type TribeMember, type HistStatsInterface } from "@/types/tribes";
 import {
   calculatePlayerGameStatsDetailed,
   calculateWinsByWeight,
