+++ components/tribes/players/PlayerSubTabs.tsx
@@ -4,9 +4,7 @@ import { useState } from "react";
 import { motion } from "motion/react";
 import { cn } from "@/lib/utils";
 import { Users, User, Swords } from "lucide-react";
-import { GameSession } from "@/components/tribes/TribeHomeTab";
-import { TribeMember } from "@/components/tribes/TribePlayersTab";
-import { HistStatsInterface } from "@/components/tribes/TribePageClient";
+import { type GameSession, type TribeMember, type HistStatsInterface } from "@/types/tribes";
 import AllPlayersView from "./AllPlayersView";
 import IndividualPlayerView from "./IndividualPlayerView";
 import HeadToHeadView from "./HeadToHeadView";
